Server/requirements.txt
anyio==4.6.0
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
Django==5.1.1
django-cors-headers==4.5.0
django-telegram-login==0.2.3
djangorestframework==3.15.2
exceptiongroup==1.2.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
packaging==24.1
pycparser==2.22
pyOpenSSL==24.2.1
python-decouple==3.8
python-dotenv==1.0.1
python-telegram-bot==21.6
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
tornado==6.4.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
